perf: narrow executor_id to u32 to keep the header at 40 bytes
Merging the community fork brought their wider waker refcount, AtomicI32
where we had AtomicI16, which pushed Header from 40 to 48 bytes: the two
extra bytes crossed an alignment boundary and cost eight.

executor_id never needed to be a usize. It is an index into the global
sleep-notifier registry and is only ever compared or looked up, so a u32
covers four billion executors and converts at the two spawn call sites
and in my_id.

Header is back to 40 bytes with their refcount included:

  awaiter          16
  vtable            8
  executor_id       4
  task_queue_index  4
  references        4
  state             1
  latency_matters   1
  end padding       2

Kept separate from the merge commit so the type change is reviewable on
its own rather than buried among conflict resolutions.

Suite green: 440 lib, 11 integration, 191 doctests, Miri clean.

+++ glommio/src/executor/mod.rs
@@ -1383,7 +1383,7 @@ impl LocalExecutor {
 
         let id = self.id;
         let ex = tq.borrow().ex.clone();
-        ex.spawn_and_run(id, tq, future)
+        ex.spawn_and_run(id as u32, tq, future)
     }
 
     /// Spawns a task directly on this executor instance.
@@ -1426,7 +1426,7 @@ impl LocalExecutor {
         let id = self.id;
 
         // can't run right away, because we need to cross into a different task queue
-        Ok(ex.spawn_and_schedule(id, tq, future))
+        Ok(ex.spawn_and_schedule(id as u32, tq, future))
     }
 
     fn preempt_timer_duration(&self) -> Duration {
